Unnecessary Files deleted
